Initial Commit
The old code really is very old; though some of it's been tweaked over
the years some of the lines haven't been touched in 11 years.  And some
of it isn't really of general interest either.

Interestingly,  I decided to split out the best parts of MyList and
polish them up,  which became data-ordlist.   And when I created
NumberSieves I came back and replaced some of the relevant parts of
old/NumberTheory.hs that were done poorly in comparison.

There is still more code here that is worth polishing up and putting on
hackage;  I've got a small start on that task in Math/NumberTheory.hs